Auto-merged by TNRP. Authorized for merge by xiao.zhou@salesforce.com.

+++ sam-internal/k8s-in/build.sh
@@ -3,6 +3,12 @@
 # Stop execution on first error
 set -ex
 
+if [ "$1" == "-h" ]; then
+  echo "Usage: build.sh - with no arguments processes all kingdoms and estates"
+  echo "       build.sh kingdom/estate1,kingdom/estate2 - processes a single estate, for local testing only!"
+  exit 1
+fi
+
 # Use this to get hypersam env var
 DIR="${BASH_SOURCE%/*}"
 if [[ ! -d "$DIR" ]]; then DIR="$PWD"; fi
@@ -34,9 +40,14 @@ if [ ! -f jsonnet/jsonnet ]; then
 fi
 
 # Nuke output folder to ensure we don't keep around stale output files
-rm -rf ../k8s-out/**
-mkdir -p ../k8s-out/
-cp k8s-out-access.yaml ../k8s-out/access.yaml
+if [ $1 -eq "" ]; then
+  rm -rf ../k8s-out/**
+  mkdir -p ../k8s-out/
+  cp k8s-out-access.yaml ../k8s-out/access.yaml
+else
+  echo "Skipping cleanup because we are processing a filtered list of estates"
+fi
+
 
 # Format input jsonnet files.  TODO: Auto-compute these directories
 for jdir in . sam sam/configs sam/templates sam/templates/rbac sdn sdn/templates slb slb/templates storage/templates flowsnake flowsnake/templates; do
@@ -49,17 +60,17 @@ else
   /opt/sam/manifestctl generate-pool-list --in ../pools/ --out  ../k8s-in/sam/configs/generated-pools.jsonnet
 fi
 
-time ./parallel_build.py sam/templates/,sdn/templates/,slb/templates/,storage/templates/ ../k8s-out/ ../pools/
-time ./parallel_build.py flowsnake/templates,sdn/templates ../k8s-out/ flowsnakeEstates.json
+./parallel_build.py --src=sam/templates/,sdn/templates/,slb/templates/,storage/templates/ --out=../k8s-out/ --pools=../pools/ --estatefilter=$1
+./parallel_build.py --src=flowsnake/templates,sdn/templates --out=../k8s-out/ --pools=flowsnakeEstates.json --estatefilter=$1
 # Skip SDN templates for Minikube
-time ./parallel_build.py flowsnake/templates ../k8s-out/ flowsnakeMinikubeEstates.json
+./parallel_build.py --src=flowsnake/templates --out=../k8s-out/ --pools=flowsnakeMinikubeEstates.json --estatefilter=$1
 
 # Skipper is a tool for creating dev/test k8s clusters in Private Cloud created
 # by the Storage Foundation team.
 # https://git.soma.salesforce.com/SFStorage/skipper
 # The line below creates k8s-out yaml files from jsonnet from storage templates
 # that are appropriate for this environment.
-time ./parallel_build.py storage/templates/ ../k8s-out/ skipperEstates.json
+./parallel_build.py --src=storage/templates/ --out=../k8s-out/ --pools=skipperEstates.json --estatefilter=$1
 
 # Json is quite poor when it comes to multi-line strings.  Since configMaps are a kubernetes resource with files
 # encoded as strings within that file, you end up with generated configMaps with enormous lines that are hard to read
